requirements.txt
--index-url https://pypi.org/simple
fastapi
uvicorn
openai
python-dotenv